Developer:  Don Lucas Bug #2541371 - No connect indicator isn't drawn properly. Modified the no connect paint method to draw a perfect 45 degree cross.  This code was previously deliberately offset by 1 pixel in the X axis for reasons unknown.
Found another occurrence of the same problem with pins marked as noconnect from within a symbol.  The code was repeated here and needed to be fixed as well.

I also fixed a similar though unreported problem when drawing pins of type "cross".  These also had a 1 pixel offset for unknown reasons.

git-svn-id: https://216.105.38.17/p/tinycad/code/TinyCAD/trunk@108 41062d70-b26d-460a-9302-f438a8930db8