Setup the Bug Tracker properties to work with the SourceForge TinyCAD bug tracker.  From here on, you will be prompted when committing a file to enter the Bug tracker number.  You can still commit without a number, but I would prefer that a number be entered each time.  Note that multiple bug numbers can be entered with a comma separated list.  No punctuation other than the comma is permitted (i.e., no spaces, semicolons, or # signs).  SVN will now automatically add a clickable link onto the end of every commit log message that is entered with a bug tracker number at the end of the message.  If you enter this each time, you may omit typing the bug number into the body of the log message.
Note that these changes are specific to the TinyCAD Bug Tracker.  You can enter an issue number from one of the other TinyCAD trackers such as the New Feature tracker, but you will get an error message that that issue has been moved to another tracker.  You will be presented with a link that will allow you to "find" the new tracker.  Unless someone can identify a better method to accomplish this, I say that this is ok behavior.  You will have to click through 2 links, but it seems to work properly although the message is mildly misleading.  This issue occurs because SVN is only set up to use a single issue tracker per directory tree and one of the setup parameters identifies the URL of the issue tracker which includes an id# for TinyCAD and an id# for the Bug tracker sub-category. Luckily, the issue/artifact ids appear to be globally unique and thus SourceForge can automatically generate a rerouted link if the wrong tracker id is used (such as using the Bug tracker id instead of using the Feature tracker id).  If you omit the tracker id, the results are not really useful.

git-svn-id: https://216.105.38.17/p/tinycad/code/TinyCAD/trunk@187 41062d70-b26d-460a-9302-f438a8930db8