Changed format string of all CDPoint strings to use %11.05f conversion.  This adds trailing zeros to the right of the decimal point so that the true intended precision of the number is represented.  It is also easier to edit the XML file when there are placeholder zeros in the string rather than extra spaces.  I would have settled for a variable number of decimal places depending on precision if I could have eliminated the long string of extra spaces, but I wasn't able to find a format specifier that would accomplish this.  There is now room for 5 decimal places, a decimal point, 4 digits of magnitude to the left of the decimal place, and an optional sign.
Part of the tradeoffs for this format is that it is used for many purposes, some of which benefit from the trailing zeros such as the coordinate readout box and some of which do not (the XML files).

At least at this point, sufficient precision is maintained up trough ANSI F size drawings (up to 10 x 10 meters) to retain good connectivity information on fine grid settings.
Issue/Tracker #2649441


git-svn-id: https://216.105.38.17/p/tinycad/code/TinyCAD/trunk@208 41062d70-b26d-460a-9302-f438a8930db8