Support for checking Spice macros of the form $(pin#) modified to first check for a pin number, and then if not found, check to see if it is a pin name.  If it is not a pin number or a pin name, then the original error message is generated. Developer:  Jason Sachs.  Note:  This code committed by Don Lucas on behalf of Jason Sachs.
While working on this problem, Jason noticed that the pin number check was being made with code that was sensitive to case.  Ball grid arrays usually use a pin "number" that includes a letter such as "a29".  Jason suggested, and I agree, that the pin number and the pin name should be made in a case insensitive way so that pin number "c29" is the same as pin number "C29".  Also, pin name "clk" will match pin name "Clk" and pin name "CLK".  Of course, net names are not affected by this convention and will continue to be case sensitive.  Presently, PSpice allows for pin numbers and pin names to be used interchangeably while straight publicly available versions of Spice (i.e., not PSpice which is a specific brand of Spice code) permit only pin numbers to be used (and only purely numeric pin numbers at that).

This coding enhancement was made primarily to support the Gandalf Speak block diagram compiler that uses pin names in its netlister almost exclusively rather than pin numbers.

Issue/Tracker #2154531


git-svn-id: https://216.105.38.17/p/tinycad/code/TinyCAD/trunk@217 41062d70-b26d-460a-9302-f438a8930db8