Updated reference designators for netlist generation to handle hierarchical designs.
I have added a get_reference_path() function to generate a hierarchical "path" (e.g. U3_H1_H4) from a symbol (CDrawMethod) and a subcircuit instance (CImportFile) along with configuration options (whether to number in forward or reverse order, and what separator to use).

This supercedes the use of GetRefSheet(), except for the fact that sheet-numbering is not yet included with get_reference_path(). Also get_reference_path is a member of CNetList rather than CDrawMethod, I suppose it should be CDrawMethod (or at least a mixin class that deals with the symbol aspects of a symbol rather than its GUI aspects)
Issue/Tracker #2174551


git-svn-id: https://216.105.38.17/p/tinycad/code/TinyCAD/branches/b_js_jason's_branch@225 41062d70-b26d-460a-9302-f438a8930db8