Feature contribution from Oleg Skydan:
The changes affected only two source files (I needed to modify
some other files to compile TinyCAD using VC2003 Std Edition,
but that is not related to the references generation):
1. stdafx.h - I needed to include <algorithm> from the STL
2. Special.cpp - All code is there (see comments below)
 
As far as I understand it is version 2.70.00 build #248, 
I have downloaded files on 04.10.2009 from the sourceforge server.
 
The stdafx.h and Special.h are attached, here are some comments 
how it works.
 
The modified code is in the CTinyCadView::OnSpecialAnotate() function.
The original code is used to generate references, but now it processes 
the sorted vector of the drawing objects. 
 
So I make the vector with all drawing objects and sort it using the 
std::sort and custom objects compare function (see DORefComp in 
Special.cpp).
 
It compares the vertical position of the objects so the top objects 
will be at the beginning of the vector and the botton ones at the end.
If two objects have the same vertical position then the leftmost ones
will appear before the rightmost ones in the sorted vector. 
 
Now the iterator of the sorted vector is supplyed to the orginal algorithm
which generates references.
 
As far as I understand the modified code is suboptimal (it will sort objects
each pass), but its current performance is acceptable (I am not familiar with
the TinyCAD code so just tried to keep the changes minimal). 
 
Feel free to ask questions regarding this changes...
 
Kind regards,
Oleg Skydan


git-svn-id: https://216.105.38.17/p/tinycad/code/TinyCAD/trunk@257 41062d70-b26d-460a-9302-f438a8930db8