For reasons that elude me, Petr Hluzin's changes to the library editor exposed some kind of bug in the ATL CStringList collection.  This caused the library load function run on startup to always fail to load the last library in the list.  The string that contained the last library in the list was valid, but the C++ concatenation operator refused to concatenate onto it for reasons that I can't figure out.  I changed code that looked something like
		CString sSearch = sLibName + _T(".idx");

to something like this:

		CString sSearch;
		sSearch.Format(_T("%s.idx"), sLibName);

to fix the problem although I don't see anything wrong with the first method.  The debugger showed that the C runtime didn't want to concatenate anyting onto the end of sLibName, but only during the last iteration of the loop when sLibName contained the last string value from the CStringList.

While studying this problem, I realized that when adding a new library to the list, newer library formats such as .TCLib are favored over older library formats such as .mdb when both libraries are present, but when the next restart of TinyCAD occurs, the startup code (the same code that had the string concatenation problem) would favor the older library formats over newer library formats as the library names are stored in the registry without type information (such as the extension).  I changed the code to favor newer formats over older formats whenever libraries are loaded at startup.  This will only affect libraries that have several versions of file formats available in the same directory.


Issue/Tracker #2370118


git-svn-id: https://216.105.38.17/p/tinycad/code/TinyCAD/trunk@350 41062d70-b26d-460a-9302-f438a8930db8