Cleaned up CRegistry::GetString() to prevent asking for doubly large buffers.  I like to be safe also, but asking for too much memory causes as many problems as asking for too little.  Also removed the argument from the ReleaseBuffer() function.  It is safer to let the function determine the amount of excess memory to release.
These changes were prompted by trying to find any other reason that a library string might be invalid.  It appears that this function is not involved in this problem, but I have decided to leave the cleanups in place.

For reference purposes, here is part of the original bug report:
For reasons that elude me, Petr Hluzin's changes to the library editor exposed some kind of bug in the ATL CStringList collection.  This caused the library load function run on startup to always fail to load the last library in the list.  The string that contained the last library in the list was valid, but the C++ concatenation operator refused to concatenate onto it for reasons that I can't figure out. 
Issue/Tracker #2370118


git-svn-id: https://216.105.38.17/p/tinycad/code/TinyCAD/trunk@351 41062d70-b26d-460a-9302-f438a8930db8