This fix was contributed by Mark Langezaal as documented in this email fragment:
Hello Don,

Solved............
Attached is the changed source file.
In the process I have actually solved three bugs:
1) Disappearing wire segments.  The root-cause was zero-length wires.
2) Tool-window state was somtimes invalid after wire dragging.  This was caused by the merge/deletion of the currently selected object, which caused this object to be unselected without a tool-window deselect.
3) Wire segment b-side is not merged when a-side is already merged.   
This is not an actual bug I could reproduce, but I deducted this from my code-read of the Merge method.
The code tested only for !merge_a  and not for !merge_b.  In any case this was wrong.  My proposed solution is to remove the test for !merge_a, because it doesn't matter if merge_a is already set or not. Of course the same applies for merge_b.

Hope you like it!
Best regards,
/Mark

Issue/Tracker #2206894,2715286


git-svn-id: https://216.105.38.17/p/tinycad/code/TinyCAD/trunk@369 41062d70-b26d-460a-9302-f438a8930db8