Updgraded TinyCAD to Visual Studio 2015 (no updates) to Visual Studio 2015 update 2.  This mostly concerned working around bugs in the Visual Studio update 2 release as well as fixing a moderate number of instances of passing a CString without conversion to any of the xxxprintf functions.  The new update 2 now complains about this and wants an appropriate character string to be passed instead.  I modified the CStrings to use the .GetBuffer() function to convert to the proper string type.
The VS2015.sln files were broken due to the bugs in update 2 so I had to change some of the options to work around the compiler bugs.  All files are fixed now, but this commit is only minimally tested.

I extended the screen space available to display the Windows version in the help/about dialog box as I now include Windows 10 options and a better string for unknown versions.  Windows 8 and Windows 10 added a large number of new options.  It also deprecated GetVersionEx() as there is a new API Microsoft wants you to use (not implemented in TinyCAD yet).  Microsoft changed the GetVersionEX() function to return the version that TinyCAD thinks is being used instead of what is actually being used due to an expanded set of compatibility options.

I also added a Visual Studio controlled manifest now controlled via the project file in order to specify that TinyCAD is not aware of high dots per inch (DPI) options.  Otherwise, the icons and dialogs are very tiny on a 15" diagonal 4K monitor such as the Dell XPS 15 laptop.  Converting to high DPI functionality is an upgrade for another day...

git-svn-id: https://216.105.38.17/p/tinycad/code/TinyCAD/branches/b_ml_vs2015_mfc_feature_pack@664 41062d70-b26d-460a-9302-f438a8930db8