Incremental improvements to the installer.  Added a "quiet mode" inhibit on the unconditional library installation path dialog box at the request of an institutional customer who likes to push the installations out to a lot of workstations.
Added a convenience link to the user's group on Yahoo Groups.

Updated the release number to 2.90.  I will let the TinyCAD help screen give the exact release number including the build number.

Modified the name of the output installer file name from setup.exe to the usual format used in our releases.  This avoids the issue that if you click on the setup.exe or use the NSIS "Test Installer" button, Windows will insert compatibility code into the installer and break it, according to NSIS.

git-svn-id: https://216.105.38.17/p/tinycad/code/TinyCAD/trunk@673 41062d70-b26d-460a-9302-f438a8930db8