Thank you Niccola Torres <niccolatorres@users.sf.net> for your solution to this bug.  It will be incorporated in release 1.90 and is in the trunk of the repository now.
The bug is titled "TC crash on creating BOM for hierarchical designs".  Niccola's explanation is as follows:


It seems that BOM Generator tries to assign the reference context of an empty parent, but tries to access an invalid index of an empty collection.

I've reproduced this on 4 computers (1 running Win7 pro, 1 running Win10 home, 2 running Win10 pro) with the #631 production release of TC. Whenever I insert a design as a symbol, the parts generator crashes the program and I lose all work done so far. The design has 12 instances of a circuit block that is identical.

Thinking that a different number would work, I tried this down to 2 identical blocks - no luck, still the crash. That's when I tried all my computers running different versions of Windows.

My workaround was to copy & paste the blocks 12 times - which is a step backwards.

Issue/Tracker #190


git-svn-id: https://216.105.38.17/p/tinycad/code/TinyCAD/trunk@682 41062d70-b26d-460a-9302-f438a8930db8